templates in their own directories

+++ vim/install.sh
@@ -2,14 +2,31 @@
 
 DIR="$(dirname "$(readlink -f "$0")")"
 
-echo "Load Vundle"
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+# Loading Vundle
+echo -e "\e[33mLoading Vundle...\e[0m"
+vundle_directory="$HOME/.vim/bundle/Vundle.vim"
+if [ -d $vundle_directory ]; then
+  echo -e "Directory already exists\n"
+else
+  git clone https://github.com/VundleVim/Vundle.vim.git $vundle_directory
 
-echo "Create vimrc.bak"
-cp $HOME/.vimrc $HOME/.vimrc.bak
+  echo -e "\e[32mSuccess\e[0m\n"
+fi
 
-echo "Copy vimrc"
+# Creating backup
+vimrc_file="$HOME/.vimrc"
+if [ -e $vimrc_file ]; then
+  echo -e "\e[33mCreating backup file (.vimrc.bak)...\e[0m"
+  cp $HOME/.vimrc $HOME/.vimrc.bak
+  echo -e "\e[32mSuccess\e[0m\n"
+fi
+
+# Copying template
+echo -e "\e[33mInstall vimrc...\e[0m"
 cp $DIR/vimrc $HOME/.vimrc
+echo -e "\e[32mSuccess\e[0m\n"
 
-echo "Install Plugins"
+# Instaliing plugins
+echo -e "\e[33mInstall Plugins...\e[0m"
 vim +PluginInstall +qall
+echo -e "\e[32mSuccess\e[0m"
